cloud_functions/process_video/requirements.txt
google-cloud-storage==2.8.0
google-auth==2.17.3
ffmpeg-python==0.2.0
requests==2.28.2
opencv-python-headless==4.7.0.72
pydantic==1.10.7
tenacity==8.2.2